--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,3 @@
+module github.com/GannettDigital/graphql
+
+go 1.12