Previously go routines were made for many fields in each request
resulting in a huge number of routines created and destroyed. This
optimizes the concurrency allowing reusing of go routines but retaining
the flexibility to add more as demand requires.

Using panics to propagate errors is bad practice and though this change
does not fix that it tries not to continue the practice. This has
resulted in a slight behavior change for non-null fields that have a
null value. This still causes an error but the partial data returned
is different. The spec does not address what partial data is expected
in the case of an error so this does not change compliance.