@tkuhlman 2019-03-29
Benchmark difference from before/after this change

name                old time/op    new time/op    delta                     
ListQuery_1-8          150µs ± 1%     146µs ± 7%     ~     (p=0.101 n=8+10) 
ListQuery_100-8        585µs ± 6%     478µs ± 4%  -18.33%  (p=0.000 n=10+10)
ListQuery_1K-8        4.15ms ±13%    3.05ms ± 4%  -26.36%  (p=0.000 n=9+9)  
ListQuery_10K-8       47.2ms ±11%    28.2ms ± 5%  -40.25%  (p=0.000 n=9+10) 
ListQuery_100K-8       291ms ±17%     182ms ± 3%  -37.33%  (p=0.000 n=9+10) 
WideQuery_1_1-8        108µs ± 2%     110µs ±16%     ~     (p=0.549 n=9+10)
WideQuery_10_1-8       254µs ± 2%     240µs ± 3%   -5.38%  (p=0.000 n=10+10)
WideQuery_100_1-8     1.64ms ± 3%    1.56ms ± 2%   -4.71%  (p=0.000 n=10+10)                                                                                  
WideQuery_1K_1-8      15.5ms ± 3%    14.5ms ± 1%   -6.60%  (p=0.000 n=10+8)
WideQuery_1_10-8       154µs ± 6%     142µs ± 4%   -7.61%  (p=0.000 n=10+10)
WideQuery_10_10-8      364µs ± 4%     312µs ± 3%  -14.22%  (p=0.000 n=10+10)
WideQuery_100_10-8    2.08ms ± 5%    1.87ms ± 2%  -10.15%  (p=0.000 n=10+10)
WideQuery_1K_10-8     19.3ms ± 7%    17.5ms ± 1%   -8.93%  (p=0.000 n=10+9)
                               
name                old alloc/op   new alloc/op   delta
ListQuery_1-8         52.7kB ± 0%    51.6kB ± 0%   -2.09%  (p=0.000 n=10+10)
ListQuery_100-8        303kB ± 0%     193kB ± 0%  -36.40%  (p=0.000 n=10+10)
ListQuery_1K-8        2.60MB ± 0%    1.49MB ± 0%  -42.73%  (p=0.000 n=10+10)
ListQuery_10K-8       26.0MB ± 1%    14.4MB ± 0%  -44.43%  (p=0.000 n=10+10)
ListQuery_100K-8       170MB ± 1%      94MB ± 0%  -44.49%  (p=0.000 n=10+10)
WideQuery_1_1-8       37.2kB ± 0%    37.0kB ± 0%   -0.48%  (p=0.000 n=10+10)
WideQuery_10_1-8      86.6kB ± 0%    83.7kB ± 0%   -3.35%  (p=0.000 n=6+10)
WideQuery_100_1-8      584kB ± 0%     554kB ± 0%   -5.19%  (p=0.000 n=10+10)
WideQuery_1K_1-8      5.69MB ± 0%    5.38MB ± 0%   -5.33%  (p=0.000 n=10+10)
WideQuery_1_10-8      49.4kB ± 0%    47.5kB ± 0%   -3.82%  (p=0.000 n=10+7)
WideQuery_10_10-8      139kB ± 0%     110kB ± 0%  -21.03%  (p=0.000 n=10+9)
WideQuery_100_10-8    1.09MB ± 0%    0.79MB ± 0%  -27.86%  (p=0.000 n=10+10)
WideQuery_1K_10-8     11.7MB ± 0%     8.6MB ± 0%  -26.05%  (p=0.000 n=10+10)
                               
name                old allocs/op  new allocs/op  delta
ListQuery_1-8          1.02k ± 0%     1.01k ± 0%   -1.85%  (p=0.000 n=10+10)
ListQuery_100-8        5.04k ± 0%     3.44k ± 0%  -31.76%  (p=0.000 n=10+10)
ListQuery_1K-8         43.3k ± 0%     27.2k ± 0%  -37.10%  (p=0.000 n=10+9)
ListQuery_10K-8         434k ± 0%      268k ± 0%  -38.27%  (p=0.000 n=10+10)
ListQuery_100K-8       2.84M ± 1%     1.75M ± 0%  -38.25%  (p=0.000 n=10+10)
WideQuery_1_1-8          663 ± 0%       656 ± 0%   -1.06%  (p=0.000 n=10+10)
WideQuery_10_1-8       1.75k ± 0%     1.70k ± 0%   -2.46%  (p=0.000 n=10+10)
WideQuery_100_1-8      12.4k ± 0%     12.0k ± 0%   -3.24%  (p=0.000 n=10+10)
WideQuery_1K_1-8        119k ± 0%      115k ± 0%   -3.36%  (p=0.000 n=10+10)
WideQuery_1_10-8         816 ± 0%       773 ± 0%   -5.27%  (p=0.000 n=10+10)
WideQuery_10_10-8      2.55k ± 0%     2.14k ± 0%  -15.79%  (p=0.000 n=10+10)
WideQuery_100_10-8     20.0k ± 0%     16.0k ± 0%  -20.00%  (p=0.000 n=8+8)
WideQuery_1K_10-8       193k ± 0%      153k ± 0%  -20.67%  (p=0.000 n=10+10)